+++ src/test/php/net/xp_framework/unittest/core/extensions/ArrayListExtensions.class.php
@@ -11,7 +11,7 @@ use lang\types\ArrayList;
 class ArrayListExtensions extends \lang\Object {
 
   static function __import($scope) {
-    \xp::extensions(__CLASS__, $scope);
+    \xp::extensions(self::class, $scope);
   }
 
   /**