+++ src/test/php/net/xp_framework/unittest/reflection/ProtectedAccessibilityFixture.class.php
@@ -1,9 +1,11 @@
 <?php namespace net\xp_framework\unittest\reflection;
 
+use lang\XPClass;
+
 /**
  * Fixture class for accessibility tests
  *
- * @see      xp://net.xp_framework.unittest.reflection.ProtectedAccessibilityTest
+ * @see   xp://net.xp_framework.unittest.reflection.ProtectedAccessibilityTest
  */
 class ProtectedAccessibilityFixture extends \lang\Object {
   protected $target= 'Target';
@@ -11,7 +13,6 @@ class ProtectedAccessibilityFixture extends \lang\Object {
 
   /**
    * Constructor
-   *
    */
   protected function __construct() { }
 
@@ -39,7 +40,7 @@ class ProtectedAccessibilityFixture extends \lang\Object {
    * @param   lang.XPClass
    * @return  net.xp_framework.unittest.reflection.ProtectedAccessibilityFixture
    */
-  public static function construct(\lang\XPClass $class) {
+  public static function construct(XPClass $class) {
     return $class->getConstructor()->newInstance([]);
   }
 
@@ -49,7 +50,7 @@ class ProtectedAccessibilityFixture extends \lang\Object {
    * @param   lang.XPClass
    * @return  string
    */
-  public static function invoke(\lang\XPClass $class) {
+  public static function invoke(XPClass $class) {
     return $class->getMethod('target')->invoke(new self());
   }
 
@@ -59,7 +60,7 @@ class ProtectedAccessibilityFixture extends \lang\Object {
    * @param   lang.XPClass
    * @return  string
    */
-  public static function invokeStatic(\lang\XPClass $class) {
+  public static function invokeStatic(XPClass $class) {
     return $class->getMethod('staticTarget')->invoke(null);
   }
 
@@ -69,7 +70,7 @@ class ProtectedAccessibilityFixture extends \lang\Object {
    * @param   lang.XPClass
    * @return  string
    */
-  public static function read(\lang\XPClass $class) {
+  public static function read(XPClass $class) {
     return $class->getField('target')->get(new self());
   }
 
@@ -79,7 +80,7 @@ class ProtectedAccessibilityFixture extends \lang\Object {
    * @param   lang.XPClass
    * @return  string
    */
-  public static function readStatic(\lang\XPClass $class) {
+  public static function readStatic(XPClass $class) {
     return $class->getField('staticTarget')->get(null);
   }
 
@@ -89,11 +90,11 @@ class ProtectedAccessibilityFixture extends \lang\Object {
    * @param   lang.XPClass
    * @return  string
    */
-  public static function write(\lang\XPClass $class) {
-    with ($s= new self(), $f= $class->getField('target')); {
-      $f->set($s, 'Modified');
-      return $f->get($s);
-    }
+  public static function write(XPClass $class) {
+    return with (new self(), $class->getField('target'), function($self, $f) {
+      $f->set($self, 'Modified');
+      return $f->get($self);
+    });
   }
 
   /**
@@ -102,10 +103,10 @@ class ProtectedAccessibilityFixture extends \lang\Object {
    * @param   lang.XPClass
    * @return  string
    */
-  public static function writeStatic(\lang\XPClass $class) {
-    with ($f= $class->getField('staticTarget')); {
+  public static function writeStatic(XPClass $class) {
+    return with ($class->getField('staticTarget'), function($f) {
       $f->set(null, 'Modified');
       return $f->get(null);
-    }
+    });
   }
 }